requirements.txt
pillow=3.0.0